Update connector.ts
when you call 'manager' to activate something via a plugin call, this call gets queued in the engine. However the iframe plugin itself also calls manager during the activation to update its profile. But that call will only return when the first one, activatePlugin has finished. So it waits, activatePugin times out and then updateProfile is executed. This only happens however when a plugin has methods. To update the profile in the manager.